requirements.txt
torch
torchvision
streamlit
Pillow
matplotlib
seaborn
pandas
numpy
scikit-learn
tqdm            